Refactor OrdersController authorization code
It basically moves authorization into each action that needs it.

In specs we were testing that passing an order :id as params to
the request it would have updated the order with that id instead
of the current_order.

This was not true, we were just authorizing actions against that
order but actually performing all actions against current order
so I think it's better to remove those specs.